Cleaned up code a lot
