+++ CoolCurrency/Extensions + Alerts/AppColours.swift
