+++ CoolCurrency/Extensions + Alerts/UserAlerts.swift
