+++ CoolCurrency/Repository/AuthenticationRepository.swift
