+++ CoolCurrency/Repository/Protocol/RepositoryProtocols.swift
