+++ CoolCurrency/ResponseModel/CurrencyDataModel.swift
