+++ CoolCurrency/ResponseModel/CurrencyResponseModel.swift
@@ -17,7 +17,6 @@ struct Response: Codable {
 }
 
 struct Rates: Codable {
-    
     let unitedStatesDollar: Double
     let euro: Double
     let indianRupee: Double
