+++ CoolCurrency/ResponseModel/UserInformation.swift
@@ -8,12 +8,12 @@
 import Foundation
 
 struct UserInformation {
-    
+
     let firstName: String
     let lastName: String
     let gender: String
     let dateOfBirth: String
-    
+
     internal init(firstName: String, lastName: String, gender: String, dateOfBirth: String) {
         self.firstName = firstName
         self.lastName = lastName
