+++ CoolCurrency/ResponseModel/en.lproj/Localizable.strings
