+++ CoolCurrency/ViewModel/ConversionCurrencyViewModel.h
