+++ CoolCurrency/ViewModel/ConversionCurrencyViewModel.m
