+++ CoolCurrency/ViewModel/Protocol/ViewModelProtocols.swift
