+++ CoolCurrency/ViewModel/SettingsViewModel.swift
@@ -9,39 +9,25 @@ import Foundation
 
 class SettingsViewModel: SettingsViewModiable {
     
+    var selectedRow = 0
+    var profilePictureDataImage: Data?
+    let currencyList = ["GBP", "ZAR", "USD", "INR", "CAD", "GHS", "JPY", "RUB", "CNY", "EUR", "AED", "BRL", "AUD"]
     private var databaseRepository: DatabaseRepositable
     private var authenticationRepository: AuthenticationRepositable
     private var userSettingsList: [String: String] = [:]
     private weak var delegate: SettingsViewModelDelegate?
     private var firstName: String?
     private var profilePictureURLString: String?
-    var profilePictureDataImage: Data?
     private var lastName: String?
     private var gender: Int?
     private var birthDate: Date?
     private var defaultCurrency: String?
     private var unitMeasurement: Int?
-    var selectedRow = 0
-    let currencyList = ["GBP", "ZAR", "USD", "INR", "CAD", "GHS", "JPY", "RUB", "CNY", "EUR", "AED", "BRL", "AUD"]
     
     init(databaseRepository: DatabaseRepositable, authenticationRepository: AuthenticationRepositable, delegate: SettingsViewModelDelegate) {
         self.databaseRepository = databaseRepository
         self.authenticationRepository = authenticationRepository
         self.delegate = delegate
-    }
-    
-    func loadUserSettingsFromDatabase() {
-        databaseRepository.retrieveUserInformationFromDatabase(userID: authenticationRepository.signedInUserIdentification(), completion: { [weak self] result in
-            do {
-                let newUserDetails = try result.get()
-                self?.userSettingsList = newUserDetails
-                self?.checkUserList()
-                self?.downloadProfileImageFromDatabase()
-                self?.delegate?.bindViewModel()
-            } catch {
-                self?.delegate?.showUserErrorMessage(error: error)
-            }
-        })
     }
     
     var retrieveFirstName: String {
@@ -72,9 +58,25 @@ class SettingsViewModel: SettingsViewModiable {
         unitMeasurement ?? 0
     }
     
+    func loadUserSettingsFromDatabase() {
+        databaseRepository.retrieveUserInformationFromDatabase(userID: authenticationRepository.signedInUserIdentification(),
+                                                               completion: { [weak self] result in
+            do {
+                let newUserDetails = try result.get()
+                self?.userSettingsList = newUserDetails
+                self?.checkUserList()
+                self?.downloadProfileImageFromDatabase()
+                self?.delegate?.bindViewModel()
+            } catch {
+                self?.delegate?.showUserErrorMessage(error: error)
+            }
+        })
+    }
+    
     func downloadProfileImageFromDatabase() {
         guard let urlString = profilePictureURLString else { return }
-        databaseRepository.performProfilePictureRequest(for: urlString, completion: { [weak self] result in
+        databaseRepository.performProfilePictureRequest(for: urlString,
+                                                           completion: { [weak self] result in
             switch result {
             case .success(let imageData):
                 self?.profilePictureDataImage = imageData
@@ -100,7 +102,8 @@ class SettingsViewModel: SettingsViewModiable {
     
     func updateFirstName(_ firstName: String) {
         databaseRepository.updateFirstNameUserInformationToDatabase(SignedInUser: authenticationRepository.signedInUserIdentification(),
-                                                                    username: firstName, completion: { [weak self] result in
+                                                                    username: firstName,
+                                                                    completion: { [weak self] result in
             switch result {
             case .success(_):
                 self?.loadUserSettingsFromDatabase()
@@ -112,7 +115,8 @@ class SettingsViewModel: SettingsViewModiable {
     
     func updateLastName(_ lastName: String) {
         databaseRepository.updateLastNameUserInformationToDatabase(SignedInUser: authenticationRepository.signedInUserIdentification(),
-                                                                   userLastName: lastName, completion: { [weak self] result in
+                                                                   userLastName: lastName,
+                                                                   completion: { [weak self] result in
             switch result {
             case .success(_):
                 self?.loadUserSettingsFromDatabase()
@@ -124,7 +128,8 @@ class SettingsViewModel: SettingsViewModiable {
     
     func updateGender(_ gender: String) {
         databaseRepository.updateUserSettingsGender(SignedInUser: authenticationRepository.signedInUserIdentification(),
-                                                    userGender: gender, completion: { [weak self] result in
+                                                    userGender: gender,
+                                                    completion: { [weak self] result in
             switch result {
             case .success(_):
                 self?.loadUserSettingsFromDatabase()
@@ -136,7 +141,8 @@ class SettingsViewModel: SettingsViewModiable {
     
     func updateDateOfBirth(_ dateOfBirth: String) {
         databaseRepository.updateUserSettingsDateOfBirth(SignedInUser: authenticationRepository.signedInUserIdentification(),
-                                                         DOB: dateOfBirth, completion: { [weak self] result in
+                                                         DOB: dateOfBirth,
+                                                         completion: { [weak self] result in
             switch result {
             case .success(_):
                 self?.loadUserSettingsFromDatabase()
@@ -148,7 +154,8 @@ class SettingsViewModel: SettingsViewModiable {
     
     func updateDefaultCurrency(_ defaultCurrency: String) {
         databaseRepository.updateDefaultCurrencyInformationToDatabase(SignedInUser: authenticationRepository.signedInUserIdentification(),
-                                                                      currency: defaultCurrency, completion: { [weak self] result in
+                                                                      currency: defaultCurrency,
+                                                                      completion: { [weak self] result in
             switch result {
             case .success(_):
                 self?.loadUserSettingsFromDatabase()
@@ -160,7 +167,8 @@ class SettingsViewModel: SettingsViewModiable {
     
     func updateMeasurementUnit(_ unit: String) {
         databaseRepository.updateMeasurementUnitToDatabase(SignedInUser: authenticationRepository.signedInUserIdentification(),
-                                                           measurementUnit: unit, completion: { [weak self] result in
+                                                           measurementUnit: unit,
+                                                           completion: { [weak self] result in
             switch result {
             case .success(_):
                 self?.loadUserSettingsFromDatabase()
@@ -168,6 +176,28 @@ class SettingsViewModel: SettingsViewModiable {
                 self?.delegate?.showUserErrorMessage(error: updateToDataError)
             }
         })
+    }
+    
+    func resetEmail(newEmail email: String) {
+        authenticationRepository.resetEmailToDatabase(newEmail: email) { [weak self] result in
+            switch result {
+            case .success(_):
+                self?.loadUserSettingsFromDatabase()
+            case .failure(let signInError):
+                self?.delegate?.showUserErrorMessage(error: signInError)
+            }
+        }
+    }
+    
+    func signOutCurrentUser() {
+        authenticationRepository.signOutUser { [weak self] result in
+            switch result {
+            case .success(_):
+                self?.delegate?.signOutBindViewModel()
+            case .failure(let signInError):
+                self?.delegate?.showUserErrorMessage(error: signInError)
+            }
+        }
     }
     
     private func checkUserList() {
@@ -191,28 +221,6 @@ class SettingsViewModel: SettingsViewModiable {
                 profilePictureURLString = value
             default:
                 break
-            }
-        }
-    }
-    
-    func resetEmail(newEmail email: String) {
-        authenticationRepository.resetEmailToDatabase(newEmail: email) { [weak self] result in
-            switch result {
-            case .success(_):
-                self?.loadUserSettingsFromDatabase()
-            case .failure(let signInError):
-                self?.delegate?.showUserErrorMessage(error: signInError)
-            }
-        }
-    }
-    
-    func signOutCurrentUser() {
-        authenticationRepository.signOutUser { [weak self] result in
-            switch result {
-            case .success(_):
-                self?.delegate?.signOutBindViewModel()
-            case .failure(let signInError):
-                self?.delegate?.showUserErrorMessage(error: signInError)
             }
         }
     }
